Adding panel for tweet languages distribution.
We store the language of tweets along with the corresponding tweet ID into the DB, and we use Grafana to visualize the distribution of tweets by language.
Also at the layer of the consumer, we only store the polarity of the tweets alongs with their ID, and we no longer store the text of the tweet.